backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[argon2]
python-jose
pandas
joblib
deepface
opencv-python-headless
numpy